refactor(goproxy): extract the shared wire-server core, dedup the MySQL/PostgreSQL brokers
The MySQL and PostgreSQL brokers carried near-identical copies of the whole wire-server
lifecycle — the accept loop, connection tracking, the bounded slot pool, the listener
lifecycle, the deadline wrapper, and the graceful drain. Only the per-connection handler
genuinely differs between them.

Extract it into two leaf packages: goproxy/drain (a Tracker — a draining signal, an in-flight
count, and a bounded wait) and goproxy/wire (a Server a protocol embeds, parameterized by a name
and its handleConn; it counts a connection in before its map insert and out after its delete, so
the drain never observes zero while one is mid-registration). mysqlproxy and pgproxy shrink to
their handler, their shutdown notice, mysql's connID, and pg's switchConn; the shared
backend-generation counter and the deadline helpers move too. Behavior is unchanged — both
brokers' full suites pass through the embedded Server.
