+++ goproxy/mysqlproxy/mysqlproxy_test.go
@@ -5,7 +5,6 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
-	"net"
 	"reflect"
 	"strings"
 	"testing"
@@ -19,22 +18,6 @@ import (
 	"github.com/ridi-oss/proxy-monster/goproxy/spi"
 	"github.com/ridi-oss/proxy-monster/mysqlwire"
 )
-
-func TestServerConnectionLimit(t *testing.T) {
-	s := New(0, spi.BackendTarget{}, nil, nil, nil)
-	for i := 0; i < maxConcurrentConnections; i++ {
-		if !s.acquireConnection() {
-			t.Fatalf("connection slot %d rejected before limit", i)
-		}
-	}
-	if s.acquireConnection() {
-		t.Fatal("connection above server-wide limit was accepted")
-	}
-	s.releaseConnection()
-	if !s.acquireConnection() {
-		t.Fatal("released connection slot was not reusable")
-	}
-}
 
 func TestNormalizeBackendOKExtractsSchemaAndStripsTracking(t *testing.T) {
 	block := mysqlwire.AppendLenencStr(nil, "other_db")
@@ -141,38 +124,6 @@ func TestRelayResultSetRejectsFragmentedColumnDefinitionForCollector(t *testing.
 	if err == nil || !strings.Contains(err.Error(), "fragmented MySQL column definitions") {
 		t.Fatalf("err=%v, want fragmented column-definition rejection", err)
 	}
-}
-
-func TestDeadlineConnBoundsIdleReadAndBlockedWrite(t *testing.T) {
-	const timeout = 100 * time.Millisecond
-
-	t.Run("read", func(t *testing.T) {
-		client, peer := net.Pipe()
-		defer client.Close()
-		defer peer.Close()
-		wrapped := withIODeadlines(client, timeout, timeout)
-		_, err := wrapped.Read(make([]byte, 1))
-		if err == nil {
-			t.Fatal("idle read did not time out")
-		}
-		if netErr, ok := err.(net.Error); !ok || !netErr.Timeout() {
-			t.Fatalf("idle read error = %v, want timeout", err)
-		}
-	})
-
-	t.Run("write", func(t *testing.T) {
-		client, peer := net.Pipe()
-		defer client.Close()
-		defer peer.Close()
-		wrapped := withIODeadlines(client, timeout, timeout)
-		_, err := wrapped.Write([]byte{0x01})
-		if err == nil {
-			t.Fatal("blocked write did not time out")
-		}
-		if netErr, ok := err.(net.Error); !ok || !netErr.Timeout() {
-			t.Fatalf("blocked write error = %v, want timeout", err)
-		}
-	})
 }
 
 // TestDialBackendCachingSHA2FullAuth proves the service-account dial completes the caching_sha2_password
